requirements.txt
streamlit
python-dateutil
psycopg2-binary
fpdf2
